pyproject.toml
[project]
name = "yapper"
version = "0.1.0"
description = ""
authors = [
    {name = "Shravan Asati",email = "dev.shravan@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-genai (>=1.20.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "yt-dlp (>=2025.6.9,<2026.0.0)",
    "srt (>=3.5.3,<4.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.172.0,<3.0.0)",
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.32.0,<1.0.0)",
    "python-multipart (>=0.0.19,<1.0.0)",
    "fastapi-cors (>=0.0.6,<0.0.7)",
    "mcp (>=1.10.1,<2.0.0)",
    "mcp-use (>=1.3.3,<2.0.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "moviepy (>=2.2.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
